DATA-006: Fix Transaction.amount Precision Issue (Store as Cents)
Issue:
- Transaction.amount stored as BigDecimal with TEXT type
- SQLite NUMERIC uses floating-point representation
- Floating-point cannot exactly represent all decimal values
- Financial calculations require exact precision (no rounding errors)

Solution:
- Convert Transaction.amount from BigDecimal to Long (cents)
- Update DataTypeConverters to multiply/divide by 100
- Migration20: Convert TEXT to INTEGER (cents)
- Update Transaction entity validation for Long type
- Update Transaction constraints for INTEGER type

Data Conversion:
- 100.50 IDR -> 10050 (stored as integer cents)
- 100.00 IDR -> 10000 (stored as integer cents)
- 0.01 IDR -> 1 (stored as integer cents)
- Maximum: 999999999.99 IDR -> 99999999999 cents

Financial Application Best Practices:
- Integer arithmetic is exact (no floating-point rounding errors)
- Faster operations than text-based storage
- Compact storage (8 bytes vs variable-length text)
- Standard practice in financial applications (Stripe, PayPal, etc.)

Files Modified:
- DataTypeConverters.kt: Updated BigDecimal <-> Long (cents) conversion
- Transaction.kt: Changed amount type from BigDecimal to Long
- Transaction.kt: Updated validation for Long (cents)
- Transaction.kt: Updated companion object create() to convert to cents
- TransactionConstraints.kt: Updated TABLE_SQL for INTEGER amount
- TransactionHistoryAdapter.kt: Convert cents to BigDecimal for display
- ReceiptGenerator.kt: Convert cents to BigDecimal for receipt
- TransactionHistoryActivityTest.kt: Fixed import and updated test values

Files Created:
- Migration20.kt: Convert amount to INTEGER (cents) - 224 lines
- Migration20Down.kt: Rollback to TEXT (decimal) - 97 lines
- Migration20Test.kt: 7 test cases

Impact: HIGH - Financial data integrity improved, exact precision using integer arithmetic, prevents floating-point rounding errors in financial calculations

+++ app/src/main/java/com/example/iurankomplek/data/constraints/TransactionConstraints.kt
@@ -37,7 +37,7 @@ object TransactionConstraints {
         CREATE TABLE ${TABLE_NAME} (
             ${Columns.ID} TEXT PRIMARY KEY NOT NULL,
             ${Columns.USER_ID} INTEGER NOT NULL,
-            ${Columns.AMOUNT} NUMERIC NOT NULL,
+            ${Columns.AMOUNT} INTEGER NOT NULL,
             ${Columns.CURRENCY} TEXT NOT NULL DEFAULT 'IDR',
             ${Columns.STATUS} TEXT NOT NULL,
             ${Columns.PAYMENT_METHOD} TEXT NOT NULL,
